Move GraphQL types, fields and mutations into the "schema" folder
